+++ src/main/java/com/example/internship/repository/BookcaseRepository.java
@@ -3,5 +3,5 @@ package com.example.internship.repository;
 import com.example.internship.dao.Bookcase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookcaseRepository extends JpaRepository<Bookcase, Integer> {
+public interface BookcaseRepository extends JpaRepository<Bookcase, Long> {
 }
