+++ src/main/java/com/example/internship/repository/ClientRepository.java
@@ -3,5 +3,5 @@ package com.example.internship.repository;
 import com.example.internship.dao.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClientRepository extends JpaRepository<Client, Integer> {
+public interface ClientRepository extends JpaRepository<Client, Long> {
 }
