+++ src/main/java/com/example/internship/repository/ShelfRepository.java
@@ -3,5 +3,5 @@ package com.example.internship.repository;
 import com.example.internship.dao.Shelf;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShelfRepository extends JpaRepository<Shelf, Integer> {
+public interface ShelfRepository extends JpaRepository<Shelf, Long> {
 }
